Generate a random branching network

Generate a random branching network

Usage

brnet(
  n_patch = 50,
  p_branch = 0.5,
  mean_env_source = 0,
  sd_env_source = 1,
  rho = 1,
  sd_env_lon = 0.1,
  mean_disturb_source = 0.9,
  sd_disturb_source = 1,
  sd_disturb_lon = 0.1,
  randomize_patch = TRUE,
  plot = FALSE,
  patch_color = "env",
  patch_label = "none",
  patch_size = 3,
  n_patch_free = FALSE,
  ...
)

Arguments

n_patch

Number of patches in a network.

p_branch

Branching probability (success probability of a geometric distribution).

mean_env_source

Mean value of environmental condition at upstream terminals.

sd_env_source

SD of environmental condition at upstream terminals.

rho

Strength of spatial autocorrelation in environmental condition.

sd_env_lon

SD of longitudinal environmental noise.

mean_disturb_source

Mean disturbance strength at headwaters. The value is assumed to represent the proportional mortality (0 - 1.0) at the patch level.

sd_disturb_source

SD of disturbance strength at headwaters. The SD is defined in a logit scale with a normal distribution.

sd_disturb_lon

SD of longitudinal noise of disturbance strength. The SD is defined in a logit scale with a normal distribution.

randomize_patch

Whether randomize patches or not. If FALSE, the function may generate a biased network with ordered patches. Default TRUE.

plot

Whether a plot should be shown or not. If FALSE, a plot of the generated network will not be shown. Default TRUE.

patch_color

Type of patch (vertex) label (either "env", "disturbance" or any color code). Default "env".

patch_label

Type of patch (vertex) label (either "patch", "branch", "n_upstream"). "patch" shows patch ID, "branch" branch ID, and "n_upstream" the number of upstream contributing patches. If "none", no label will be shown on patches in the plot. Default "none".

patch_size

Patch (vertex) size in the plot.

n_patch_free

Whether imposing a constraint on n_patch. If TRUE, the number of patches a random variable following a negative binomial distribution.

...

Arguments passed to ggraph::geom_node_label().

Reference

see package webpage for instruction

Author

Akira Terui, hanabi0111@gmail.com

Examples

if (FALSE) {
# not run
brnet(n_patch = 10, p_branch = 0.5)
}