Simulate metacommunity dynamics

Simulate metacommunity dynamics

mcsim(
  n_species = 5,
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  propagule_interval = NULL,
  carrying_capacity = 100,
  interaction_type = "constant",
  alpha = 0,
  min_alpha = NULL,
  max_alpha = NULL,
  r0 = 4,
  niche_optim = NULL,
  min_optim = -1,
  max_optim = 1,
  sd_niche_width = NULL,
  min_niche_width = 0.1,
  max_niche_width = 1,
  niche_cost = 1,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL,
  landscape_size = 10,
  mean_env = 0,
  sd_env = 0.1,
  spatial_env_cor = FALSE,
  phi = 1,
  p_dispersal = 0.1,
  theta = 1,
  plot = FALSE
)

Arguments

n_species

numeric value. Number of species in a metacommunity.

n_patch

numeric value. Number of patches in a metacommunity.

n_warmup

numeric value. Number of time-steps for warm-up. Default 200.

n_burnin

numeric value. Number of time-steps for burn-in. Default 200.

n_timestep

numeric value. Number of time-steps to be saved. Default 1000.

propagule_interval

numeric value. Time interval for propagule introduction during warm-up. If NULL, a value of ceiling(n_warmup / 10) will be used.

carrying_capacity

numeric value (length should be one or equal to n_patch). Carrying capacities of individual patches. Default 100.

interaction_type

character string. "constant" or "random". "constant" assumes the unique interaction strength of alpha for all pairs of species. "random" draws random numbers from a uniform distribution with min_alpha and max_alpha.

alpha

numeric value. Species interaction strength. Enabled if interaction_type = "constant". Default 0.

min_alpha

numeric value. Minimum value of a uniform distribution that generates species interaction strength. Enabled if interaction_type = "random". Default NULL.

max_alpha

numeric value. Maximum value of a uniform distribution that generates species interaction strength. Enabled if interaction_type = "random". Default NULL.

r0

numeric value (length should be one or equal to n_species). Maximum reproductive number of the Beverton-Holt model.

niche_optim

numeric value (length should be one or equal to n_species). Niche optimum of species (environmental value that maximizes the reproductive number). Default NULL.

min_optim

numeric value. Minimum value of a uniform distribution that generates optimal environmental values of simulated species. Values are randomly assigned to species. Enabled if niche_optim = NULL.

max_optim

numeric value. Maximum value of a uniform distribution that generates optimal environmental values of simulated species. Values are randomly assigned to species. Enabled if niche_optim = NULL.

sd_niche_width

numeric value (length should be one or equal to n_species). Niche width of species. Higher values indicate greater niche width.

min_niche_width

numeric value. Minimum value of a uniform distribution that generates niche width values of simulated species. Values are randomly assigned to species. Enabled if sd_niche_width = NULL.

max_niche_width

numeric value. Maximum value of a uniform distribution that generates niche width values of simulated species. Values are randomly assigned to species. Enabled if sd_niche_width = NULL.

niche_cost

numeric value. Determine the cost of wide niche (smaller values imply greater costs of wider niche). Default 1.

xy_coord

data frame. Each row should correspond to an individual patch, with x and y coordinates (columns). Defualt NULL.

distance_matrix

numeric value. Distance matrix indicating distance between habitat patches. If provided, the distance matrix will be used to generate dispersal matrix and to calculate distance decay of environmental correlations. Default NULL.

dispersal_matrix

numeric value. Dispersal matrix to be used to simulate dispersal process. Override distance_matrix. Default NULL.

landscape_size

numeric value. Length of a landscape on a side. Enabled if dispersal_matrix = NULL.

mean_env

numeric value (length should be one or equal to n_patch). Mean environmental values of patches.

sd_env

numeric value. Standard deviation of temporal environmental variation at each patch.

spatial_env_cor

logical. If TRUE, spatial autocorrelation in temporal environmental fluctuation is considered. Default FALSE.

phi

numeric value. A parameter describing the distance decay of spatial autocorrelation in temporal environmental fluctuation. Enabled if spatial_env_cor = TRUE.

p_dispersal

numeric value (length should be one or equal to n_species). Probability of dispersal.

theta

numeric value. Dispersal parameter describing dispersal capability of species.

plot

logical. If TRUE, five sample patches and species of df_dynamics are plotted.

Value

df_dynamics data frame containing simulated metacommunity dynamics. df_species data frame containing species attributes. df_patch data frame containing patch attributes. df_diversity data frame containing diversity metrics. alpha and gamma diversities were averaged over the simulation time steps (n_timestep). beta diversity is calculated as gamma / alpha.

Reference

see github page for instruction

Author

Akira Terui, hanabi0111@gmail.com

Examples

if (FALSE) {
# not run
mcsim(n_patch = 5, n_species = 5)
}