Generate a random branching network

Generate a random branching network

brnet(
  n_patch = 50,
  p_branch = 0.5,
  mean_env_source = 0,
  sd_env_source = 1,
  rho = 1,
  sd_env_lon = 0.1,
  mean_disturb_source = 0.9,
  sd_disturb_source = 1,
  sd_disturb_lon = 0.1,
  randomize_patch = TRUE,
  plot = TRUE,
  patch_label = NULL,
  patch_size = 6,
  patch_scaling = TRUE,
  scale_factor = 8,
  n_patch_free = FALSE
)

Arguments

n_patch

numeric value indicating the number of patches in a network.

p_branch

numeric value indicating the branching probability (success probability of a geometric distribution).

mean_env_source

numeric value indicating the mean value of environmental condition at upstream terminals.

sd_env_source

numeric value indicating the SD of environmental condition at upstream terminals.

rho

numeric value indicating the strength of spatial autocorrelation in environmental condition. The environmental condition at patch x \(z\)x is determined as \(z\)x\( = \rho\)zx-1\( + \epsilon\)x, where \(\epsilon\)x is the random variable drawn from a normal distribution with mean 0 and SD \(\sigma\)env. See github page for further details.

sd_env_lon

numeric value indicating the SD of longitudinal environmental noise.

mean_disturb_source

numeric value indicating the mean of disturbance strength at headwaters. The value is assumed to represent the proportional mortality (0 - 1.0) at the patch level.

sd_disturb_source

numeric value indicating the SD of disturbance strength at headwaters. The SD is defined in a logit scale with a normal distribution.

sd_disturb_lon

numeric value indicating the SD of longitudinal noise of disturbance strength. The SD is defined in a logit scale with a normal distribution.

randomize_patch

logical indicating whether randomize patches or not. If FALSE, the function may generate a biased network with ordered patches. Default TRUE.

plot

logical indicating if a plot should be shown or not. If FALSE, a plot of the generated network will not be shown. Default TRUE.

patch_label

character string indicating a type of patch (vertex) label (either "patch", "branch", "n_upstream"). "patch" shows patch ID, "branch" branch ID, and "n_upstream" the number of upstream contributing patches. If NULL, no label will be shown on patches in the plot. Default NULL.

patch_size

patch (vertex) size in the plot. Default 6.

patch_scaling

logical. If TRUE, patch (vertex) size will be proportional to the number of upstream contributing patches. The patch (vertex) size will be equal to 0.3 * scale_factor at the upstream terminals and 1.3 * scale_factor at the root. Overrides patch_size.

scale_factor

numeric value scaling patch (vertex) size. Enabled if patch_scaling = TRUE.

n_patch_free

logical value indicating whether imposing a constraint on n_patch. If TRUE, the number of patches a random variable following a negative binomial distribution.

Value

adjacency_matrix adjacency matrix for the generated network. distance_matrix distance matrix for the generated network. df_patch data frame containing patch attributes.

Reference

see github page for instruction

Author

Akira Terui, hanabi0111@gmail.com

Examples

if (FALSE) {
# not run
brnet(n_patch = 10, p_branch = 0.5)
}