Generate a random branching network

Generate a random branching network

Usage

brnet(
  n_patch = 50,
  p_branch = 0.5,
  mean_env_source = 0,
  sd_env_source = 1,
  rho = 1,
  sd_env_lon = 0.1,
  mean_disturb_source = 0.9,
  sd_disturb_source = 1,
  sd_disturb_lon = 0.1,
  randomize_patch = TRUE,
  n_patch_free = FALSE
)

Arguments

n_patch

Number of patches in a network.

p_branch

Branching probability (success probability of a geometric distribution).

mean_env_source

Mean value of environmental condition at upstream terminals.

sd_env_source

SD of environmental condition at upstream terminals.

rho

Strength of spatial autocorrelation in environmental condition.

sd_env_lon

SD of longitudinal environmental noise.

mean_disturb_source

Mean disturbance strength at headwaters. The value is assumed to represent the proportional mortality (0 - 1.0) at the patch level.

sd_disturb_source

SD of disturbance strength at headwaters. The SD is defined in a logit scale with a normal distribution.

sd_disturb_lon

SD of longitudinal noise of disturbance strength. The SD is defined in a logit scale with a normal distribution.

randomize_patch

Whether randomize patches or not. If FALSE, the function may generate a biased network with ordered patches. Default TRUE.

n_patch_free

Whether imposing a constraint on n_patch. If TRUE, the number of patches a random variable following a negative binomial distribution.

Reference

see package webpage for instruction

Author

Akira Terui, hanabi0111@gmail.com

Examples

if (FALSE) {
# not run
brnet(n_patch = 10, p_branch = 0.5)
}