Simulate meta-food web dynamics with intraguild predation

Simulate meta-food web dynamics with intraguild predation

Usage

igpsim(
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  r_b = 10,
  conv_eff = 0.9,
  attack_rate = 0.05,
  handling_time = 0.5,
  s = 0,
  propagule_interval = NULL,
  propagule_seed = c(100, 10, 1),
  carrying_capacity = 100,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL,
  p_disturb = 0,
  i_disturb = 0,
  phi_disturb = Inf,
  landscape_size = 10,
  p_dispersal = 0.1,
  theta = 1,
  plot = FALSE
)

Arguments

n_patch

Number of patches in a metacommunity.

n_warmup

Number of time-steps for warm-up. Default 200.

n_burnin

Number of time-steps for burn-in. Default 200.

n_timestep

Number of time-steps to be saved. Default 1000.

r_b

Maximum reproductive rate of basal species

conv_eff

Energetic conversion efficiency; must be given by the order of basal to ig-prey, basal to ig-predator, and ig-prey to ig-predator

attack_rate

Attack rate. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator

handling_time

Handling time. Must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator

s

Strength of switching between basal and ig-prey (confined to 0 - 1). Switching to basal species is more likely to occur with higher values.

propagule_interval

Time interval for propagule introduction during warm-up. If NULL, a value of ceiling(n_warmup / 10) will be used.

propagule_seed

Propagule mean density (intensity parameter in a Poisson distribution). Should be given as a scalar or vector. If given as a vector, the elements should appear in order of basal, intra-guild prey, and intra-guild predator.

carrying_capacity

Carrying capacities of individual patches. Length must be one or equal to n_patch. Default 100.

xy_coord

Data frame for site coordinates. Each row should correspond to an individual patch, with x and y coordinates (columns). Defualt NULL.

distance_matrix

Distance matrix indicating distance between habitat patches. If provided, the distance matrix will be used to generate dispersal matrix and to calculate distance decay of environmental correlations. Default NULL.

dispersal_matrix

Dispersal matrix to be used to simulate dispersal process. Override distance_matrix. Default NULL.

p_disturb

Disturbance probability.

i_disturb

Disturbance-induced proportional mortality.

phi_disturb

Temporal precision of disturbance-induced proportional mortality. Set Inf to assume no temporal variability.

landscape_size

Length of a landscape on a side. Enabled if dispersal_matrix = NULL.

p_dispersal

Probability of dispersal. Length must be one or equal to n_species.

theta

Dispersal parameter describing dispersal capability of species.

plot

If TRUE, five sample patches and species of df_dynamics are plotted.

Reference

see package webpage for instruction

Author

Akira Terui, hanabi0111@gmail.com