Simulate meta-food web dynamics with intraguild predation

Simulate meta-food web dynamics with intraguild predation

igpsim(
  n_patch = 5,
  n_warmup = 200,
  n_burnin = 200,
  n_timestep = 1000,
  r_b = 5,
  conv_eff = rep(5, 3),
  attack_rate = rep(0.5, 3),
  handling_time = rep(1, 3),
  s0 = rep(0.8, 3),
  propagule_interval = NULL,
  carrying_capacity = 100,
  xy_coord = NULL,
  distance_matrix = NULL,
  dispersal_matrix = NULL,
  p_disturb = 0,
  m_disturb = 0,
  landscape_size = 10,
  p_dispersal = 0.1,
  theta = 1,
  plot = FALSE
)

Arguments

n_patch

numeric value. Number of patches in a metacommunity.

n_warmup

numeric value. Number of time-steps for warm-up. Default 200.

n_burnin

numeric value. Number of time-steps for burn-in. Default 200.

n_timestep

numeric value. Number of time-steps to be saved. Default 1000.

r_b

maximum reproductive rate of basal species

conv_eff

energetic conversion efficiency; must be given by the order of basal to ig-prey, basal to ig-predator, and ig-prey to ig-predator

attack_rate

attack rate; must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator

handling_time

handling time; ; must be given by the order of basal to ig_prey, basal to ig-predator, and ig-prey to ig-predator

s0

background survival rate; must be given by the order of basal, ig-prey, and ig-predator

propagule_interval

numeric value. Time interval for propagule introduction during warm-up. If NULL, a value of ceiling(n_warmup / 10) will be used.

carrying_capacity

numeric value (length should be one or equal to n_patch). Carrying capacities of individual patches. Default 100.

xy_coord

data frame. Each row should correspond to an individual patch, with x and y coordinates (columns). Defualt NULL.

distance_matrix

numeric value. Distance matrix indicating distance between habitat patches. If provided, the distance matrix will be used to generate dispersal matrix and to calculate distance decay of environmental correlations. Default NULL.

dispersal_matrix

numeric value. Dispersal matrix to be used to simulate dispersal process. Override distance_matrix. Default NULL.

p_disturb

disturbance probability.

m_disturb

disturbance-induced proportional mortality.

landscape_size

numeric value. Length of a landscape on a side. Enabled if dispersal_matrix = NULL.

p_dispersal

numeric value (length should be one or equal to n_species). Probability of dispersal.

theta

numeric value. Dispersal parameter describing dispersal capability of species.

plot

logical. If TRUE, five sample patches and species of df_dynamics are plotted.

Value

df_dynamics data frame containing simulated metacommunity dynamics. df_species data frame containing species attributes. df_patch data frame containing patch attributes. df_int parameters for trophic interactions

Reference

see github page for instruction

Author

Akira Terui, hanabi0111@gmail.com